Code cleanup and improved structure
