+++ backend/src/controllers/auth.controller.js
@@ -56,6 +56,8 @@ export const registerController = asyncHandler(async (req, res) => {
     user._id = undefined;
     user.emailVerificationToken = undefined;
     user.emailVerificationExpires = undefined;
+    user.createdAt = undefined;
+    user.updatedAt = undefined;
 
     return res
         .cookie("accessToken", accessToken, {
@@ -67,83 +69,6 @@ export const registerController = asyncHandler(async (req, res) => {
         .status(201)
         .json(new ApiResponse(201, "User created successfully. Please check your email to verify your account.", { user, accessToken }))
 })
-
-export const verifyUserMailController = asyncHandler(async (req, res) => {
-    const { token } = req.params;
-
-    if (!token) {
-        throw new ApiError(400, "Verification token is required");
-    }
-
-    // Find user with this verification token
-    const user = await User.findOne({
-        emailVerificationToken: token,
-        emailVerificationExpires: { $gt: Date.now() }
-    });
-
-    if (!user) {
-        throw new ApiError(400, "Invalid or expired verification token");
-    }
-
-    // Mark user as verified and clear verification token
-    user.isVerified = true;
-    user.emailVerificationToken = undefined;
-    user.emailVerificationExpires = undefined;
-
-    await user.save();
-
-    // Generate a new access token and set it as cookie
-    const accessToken = user.generateAccessToken();
-
-    return res
-        .status(200)
-        .cookie("accessToken", accessToken, {
-            httpOnly: true,
-            secure: process.env.NODE_ENV === "production",
-            sameSite: "strict",
-            maxAge: 7 * 24 * 60 * 60 * 1000 // 7 days
-        })
-        .json(
-            new ApiResponse(200, "Email verified successfully. You can now log in.", { accessToken })
-        );
-})
-
-export const resendVerificationEmail = asyncHandler(async (req, res) => {
-
-    const email = req.user.email
-
-    if (!email) {
-        throw new ApiError(400, "Email is required");
-    }
-
-
-    const user = await User.findOne({ email });
-
-    if (!user) {
-        throw new ApiError(404, "User not found");
-    }
-
-    if (user.isVerified) {
-        throw new ApiError(400, "Email is already verified");
-    }
-
-    // Generate new verification token
-    const verificationToken = crypto.randomBytes(32).toString('hex');
-    const verificationExpires = new Date(Date.now() + 24 * 60 * 60 * 1000); // 24 hours
-
-    // Update user with new verification token
-    user.emailVerificationToken = verificationToken;
-    user.emailVerificationExpires = verificationExpires;
-    await user.save();
-
-    // Send verification email
-    await sendVerificationEmail(user.email, verificationToken, user.fullName);
-
-    return res.status(200).json(
-        new ApiResponse(200, "Verification email resent successfully. Please check your email.")
-    );
-})
-
 
 export const loginController = asyncHandler(async (req, res) => {
     const { username, password } = req.body;
@@ -173,6 +98,8 @@ export const loginController = asyncHandler(async (req, res) => {
     userObj._id = undefined;
     userObj.emailVerificationToken = undefined;
     userObj.emailVerificationExpires = undefined;
+    userObj.createdAt = undefined;
+    userObj.updatedAt = undefined;
 
     let message = "User logged in successfully";
     if (!user.isVerified) {
@@ -195,4 +122,22 @@ export const logoutController = asyncHandler(async (req, res) => {
     return res.status(200).json(new ApiResponse(200, "User logged out successfully"))
 })
 
+export const getCurrentUserController = asyncHandler(async (req, res) => {
+    const user = await User.findById(req.user.id);
 
+    if (!user) {
+        throw new ApiError(404, "User not found");
+    }
+
+    // Remove sensitive data before sending response
+    const userObj = user.toObject();
+    userObj.password = undefined;
+    userObj.__v = undefined;
+    userObj._id = undefined;
+    userObj.emailVerificationToken = undefined;
+    userObj.emailVerificationExpires = undefined;
+    userObj.createdAt = undefined;
+    userObj.updatedAt = undefined;
+
+    return res.status(200).json(new ApiResponse(200, "User fetched successfully", { user: userObj }));
+})
