+++ backend/src/controllers/user.controller.js
@@ -1,0 +1,83 @@
+import { ApiError } from "../utils/ApiError.js";
+import { ApiResponse } from "../utils/ApiResponse.js";
+import { asyncHandler } from "../utils/asyncHandler.js";
+import crypto from "crypto";
+import User from "../models/user.model.js";
+import { sendVerificationEmail } from "../services/emailService.js";
+
+
+export const verifyUserMailController = asyncHandler(async (req, res) => {
+    const { token } = req.params;
+
+    if (!token) {
+        throw new ApiError(400, "Verification token is required");
+    }
+
+    // Find user with this verification token
+    const user = await User.findOne({
+        emailVerificationToken: token,
+        emailVerificationExpires: { $gt: Date.now() }
+    });
+
+    if (!user) {
+        throw new ApiError(400, "Invalid or expired verification token");
+    }
+
+    // Mark user as verified and clear verification token
+    user.isVerified = true;
+    user.emailVerificationToken = undefined;
+    user.emailVerificationExpires = undefined;
+
+    await user.save();
+
+    // Generate a new access token and set it as cookie
+    const accessToken = user.generateAccessToken();
+
+    return res
+        .status(200)
+        .cookie("accessToken", accessToken, {
+            httpOnly: true,
+            secure: process.env.NODE_ENV === "production",
+            sameSite: "strict",
+            maxAge: 7 * 24 * 60 * 60 * 1000 // 7 days
+        })
+        .json(
+            new ApiResponse(200, "Email verified successfully. You can now log in.", { accessToken })
+        );
+})
+
+export const resendVerificationEmail = asyncHandler(async (req, res) => {
+
+    const email = req.user.email
+
+    if (!email) {
+        throw new ApiError(400, "Email is required");
+    }
+
+
+    const user = await User.findOne({ email });
+
+    if (!user) {
+        throw new ApiError(404, "User not found");
+    }
+
+    if (user.isVerified) {
+        throw new ApiError(400, "Email is already verified");
+    }
+
+    // Generate new verification token
+    const verificationToken = crypto.randomBytes(32).toString('hex');
+    const verificationExpires = new Date(Date.now() + 24 * 60 * 60 * 1000); // 24 hours
+
+    // Update user with new verification token
+    user.emailVerificationToken = verificationToken;
+    user.emailVerificationExpires = verificationExpires;
+    await user.save();
+
+    // Send verification email
+    await sendVerificationEmail(user.email, verificationToken, user.fullName);
+
+    return res.status(200).json(
+        new ApiResponse(200, "Verification email resent successfully. Please check your email.")
+    );
+})
